LeetCode Premium questions

+++ src/main/java/com/coding/practice/LeetCode_347.java
@@ -41,7 +41,7 @@ public class LeetCode_347 {
         }
 
         for (int i = 1; i <= k; i++) {
-            result.add((Integer) heap.remove().getKey());
+            result.add(heap.remove().getKey());
         }
         return result;
     }
